formi-chatbot-backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.1
openai==1.12.0
aiofiles==23.2.1
PyPDF2==3.0.1
google-auth==2.27.0
google-api-python-client==2.118.0
python-multipart==0.0.9
websockets==12.0 